backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
python-dotenv
playwright
google-auth-oauthlib
google-api-python-client
supabase
